visionguard-backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic-settings==2.1.0

# Database
sqlmodel==0.0.14
psycopg2-binary==2.9.9

# AI & Computer Vision
ultralytics==8.1.0
opencv-python-headless==4.8.1.78
numpy==1.26.2

# Communication
fastapi-mqtt==1.1.0
gmqtt==0.6.12

# Utilities
python-dotenv==1.0.0
httpx==0.25.1